Use uint40 for tokenId

+++ src/interfaces/ICheckIn.sol
@@ -12,7 +12,7 @@ interface ICheckIn {
     /// @param _ticketId The ID of the ticket to check in
     /// @param _ticketOwner The owner of the ticket
     /// @param _tokenId The token ID of the ticket
-    function checkIn(uint64 _ticketId, address _ticketOwner, uint256 _tokenId) external;
+    function checkIn(uint64 _ticketId, address _ticketOwner, uint40 _tokenId) external;
 
     /// @notice Adds ticket admins to a ticket
     /// @param _ticketId The ID of the ticket to add admins to
