+++ src/libs/errors/CheckInErrors.sol
@@ -3,7 +3,7 @@ pragma solidity 0.8.30;
 
 error TicketUsePeriodNotStarted();
 error TicketUsePeriodHasEnded();
-error NotTicketOwner(uint256);
+error NotTicketOwner(uint40);
 error AlreadyCheckedInForDay(uint8);
 error NoAdmins();
 error AddressZeroAdmin();
